Move com.oracle.svm.configure.json from com.oracle.svm.configure to com.oracle.svm.core into com.oracle.svm.core.util.json
