+++ substratevm/src/com.oracle.svm.configure/src/com/oracle/svm/configure/config/ConfigurationMethod.java
@@ -28,8 +28,8 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Objects;
 
-import com.oracle.svm.configure.json.JsonPrintable;
-import com.oracle.svm.configure.json.JsonWriter;
+import com.oracle.svm.core.util.json.JsonPrintable;
+import com.oracle.svm.core.util.json.JsonWriter;
 
 import jdk.vm.ci.meta.MetaUtil;
