+++ substratevm/src/com.oracle.svm.configure/src/com/oracle/svm/configure/config/ConfigurationPredefinedClass.java
@@ -27,8 +27,8 @@ package com.oracle.svm.configure.config;
 
 import java.io.IOException;
 
-import com.oracle.svm.configure.json.JsonPrintable;
-import com.oracle.svm.configure.json.JsonWriter;
+import com.oracle.svm.core.util.json.JsonPrintable;
+import com.oracle.svm.core.util.json.JsonWriter;
 
 public final class ConfigurationPredefinedClass implements JsonPrintable {
