+++ substratevm/src/com.oracle.svm.configure/src/com/oracle/svm/configure/config/conditional/HumanReadableConfigurationWithOrigins.java
@@ -29,8 +29,8 @@ import java.io.StringWriter;
 import java.util.List;
 import java.util.Set;
 
-import com.oracle.svm.configure.json.JsonPrintable;
-import com.oracle.svm.configure.json.JsonWriter;
+import com.oracle.svm.core.util.json.JsonPrintable;
+import com.oracle.svm.core.util.json.JsonWriter;
 import com.oracle.svm.core.configure.ConfigurationFile;
 
 public class HumanReadableConfigurationWithOrigins implements JsonPrintable {
