+++ substratevm/src/com.oracle.svm.configure/src/com/oracle/svm/configure/filters/ComplexFilter.java
@@ -28,7 +28,7 @@ import java.io.IOException;
 
 import org.graalvm.collections.EconomicMap;
 
-import com.oracle.svm.configure.json.JsonWriter;
+import com.oracle.svm.core.util.json.JsonWriter;
 
 public class ComplexFilter implements ConfigurationFilter {
     private HierarchyFilterNode hierarchyFilterNode;
