+++ substratevm/src/com.oracle.svm.configure/src/com/oracle/svm/configure/filters/RegexFilter.java
@@ -34,7 +34,7 @@ import java.util.regex.Pattern;
 
 import org.graalvm.collections.EconomicMap;
 
-import com.oracle.svm.configure.json.JsonWriter;
+import com.oracle.svm.core.util.json.JsonWriter;
 
 public class RegexFilter implements ConfigurationFilter {
